requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-appconf==1.0.1
django-autoslug==1.8.0
django-geoposition==0.2.2
django-sticky-uploads==0.4.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
image==1.4.0
oauthlib==1.0.1
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.0
python-social-auth==0.2.12
python3-openid==3.0.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
wheel==0.24.0
